+++ bocoel/common/batches.py
@@ -1,0 +1,11 @@
+from typing import Protocol
+
+
+class Batched(Protocol):
+    @property
+    def batch_size(self) -> int:
+        """
+        The batch size used for processing.
+        """
+
+        ...